refactor(graph-gateway): use origin header extractor
