+++ graph-gateway/Cargo.toml
@@ -7,7 +7,7 @@ version = "17.0.0"
 alloy-primitives.workspace = true
 alloy-sol-types = "0.6.0"
 anyhow.workspace = true
-axum.workspace = true
+axum = { workspace = true, features = ["headers"] }
 chrono = { version = "0.4", default-features = false, features = ["clock"] }
 cost-model = { git = "https://github.com/graphprotocol/agora", rev = "3367d76" }
 eventuals = "0.6.7"
